Screenshots to document the test of data transition between on-premises and Azure environments in the NetMaze Explorer project.
This commit includes screenshots capturing the process and results of testing data transition between on-premises and Azure environments. Please note that the screenshot names might not be displayed properly in the directory. However, if there is any confusion, all these screenshots are also available in the detailed manual created during the project.