Screenshots to document the test of accessing a blocked port on WebAppVM from both authorized and unauthorized sources in the NetMaze Explorer project.
This commit includes screenshots capturing the process and results of testing access to a blocked port on WebAppVM from both authorized (AdminVM) and unauthorized (OnPremVM) sources. Please note that the screenshot names might not be displayed properly in the directory. However, if there is any confusion, all these screenshots are also available in the detailed manual created during the project. The original file names were "Test Access to a Blocked Port from an Unauthorized Source (OnPremVM)" and "Test Access to a Blocked Port from an Authorized Source (AdminVM).