Screenshots to document the NSG Logs Monitoring test conducted in the NetMaze Explorer project, focusing on capturing and analyzing NSG flow logs in Azure.
This commit includes screenshots related to the NSG Logs Monitoring test, where the flow logs were configured and monitored to verify the enforcement of a specific security rule within the WebAppNSG in the project. The test involved denying HTTP access to WebAppVM and verifying this action through the NSG logs stored in a newly created storage account. The analysis of the flow logs confirmed the successful application of the security rule. Please note that all these screenshots are also available in the detailed manual created during the project.